+++ source/src/main/java/nl/aerius/fileserver/storage/StorageService.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see http://www.gnu.org/licenses/.
  */
-package nl.overheid.aerius.fileserver.storage;
+package nl.aerius.fileserver.storage;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
