+++ source/src/main/java/nl/aerius/fileserver/util/FilenameUtil.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see http://www.gnu.org/licenses/.
  */
-package nl.overheid.aerius.fileserver.util;
+package nl.aerius.fileserver.util;
 
 import java.io.IOException;
 import java.util.regex.Pattern;
